[REFACTOR] Moved logic specific to NYTimes to it's provider. Work in progress for #7
