+++ core-lib/src/main/java/info/hossainkhan/android/core/newsprovider/NyTimesNewsProvider.java
@@ -24,13 +24,34 @@
 
 package info.hossainkhan.android.core.newsprovider;
 
+import android.content.Context;
+
+import com.google.firebase.crash.FirebaseCrash;
+
+import java.util.ArrayList;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
 
+import info.hossainkhan.android.core.CoreApplication;
+import info.hossainkhan.android.core.data.CategoryNameResolver;
+import info.hossainkhan.android.core.model.CardItem;
+import info.hossainkhan.android.core.model.NavigationRow;
 import info.hossainkhan.android.core.model.NewsProvider;
 import info.hossainkhan.android.core.model.NewsSource;
+import io.swagger.client.ApiClient;
+import io.swagger.client.api.ConsumptionFormat;
+import io.swagger.client.api.StoriesApi;
+import io.swagger.client.model.Article;
 import io.swagger.client.model.ArticleCategory;
+import io.swagger.client.model.InlineResponse200;
+import rx.Observable;
+import rx.android.schedulers.AndroidSchedulers;
+import rx.functions.Action1;
+import rx.functions.Func1;
+import rx.schedulers.Schedulers;
+import timber.log.Timber;
 
 
 /**
@@ -102,5 +123,86 @@ public final class NyTimesNewsProvider implements NewsProvider {
         categories.add(ArticleCategory.sports);
 
         return categories;
+    }
+
+    @Override
+    public Observable<List<NavigationRow>> getNewsObservable() {
+        final Context mContext = CoreApplication.getAppComponent().getContext();
+        ApiClient apiClient = CoreApplication.getAppComponent().getApiClient();
+        StoriesApi service = apiClient.createService(StoriesApi.class);
+
+        final List<ArticleCategory> categories = new ArrayList<>(CategoryNameResolver.getPreferredCategories(mContext));
+
+        int sectionSize = categories.size();
+        List<Observable<InlineResponse200>> observableList = new ArrayList<>(sectionSize);
+
+        Timber.i("Loading categories: %s", categories);
+
+        // NOTE: Unable to use java8 lambda using jack. Error: Library projects cannot enable Jack (Java 8).
+        // ASOP Issue # https://code.google.com/p/android/issues/detail?id=211386
+        Observable.from(categories).subscribe(new Action1<ArticleCategory>() {
+            @Override
+            public void call(ArticleCategory articleCategory) {
+                observableList.add(service.sectionFormatGet(articleCategory.name(), ConsumptionFormat.json.name(), null));
+            }
+        });
+
+
+
+        return Observable.merge(observableList)
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .toList()
+                .single()
+                .map(new Func1<List<InlineResponse200>, List<NavigationRow>>() {
+                    @Override
+                    public List<NavigationRow> call(final List<InlineResponse200> inlineResponse200s) {
+                        int totalResponseItemSize = inlineResponse200s.size();
+                        Timber.i("Got total responses: %d", totalResponseItemSize);
+
+                        if (totalResponseItemSize != sectionSize) {
+                            // Error
+                            FirebaseCrash.log("Unable to get all responses.");
+                        } else {
+                            List<NavigationRow> navigationHeadlines = new ArrayList<>(totalResponseItemSize+1);
+                            navigationHeadlines.add(NavigationRow.builder()
+                                    .title(mNewsSource.name())
+                                    .type(NavigationRow.TYPE_SECTION_HEADER)
+                                    .build());
+
+                            for (int i = 0; i < totalResponseItemSize; i++) {
+                                ArticleCategory articleCategory = categories.get(i);
+                                navigationHeadlines.add(
+                                        NavigationRow.builder()
+                                                .title(mContext.getString(CategoryNameResolver
+                                                        .resolveCategoryResId(articleCategory)))
+                                                .category(articleCategory)
+                                                .cards(convertArticleToCardItems(inlineResponse200s.get(i).getResults()))
+                                                .build()
+                                );
+                            }
+                            return navigationHeadlines;
+                        }
+                        return null;
+                    }
+                });
+    }
+
+
+    /**
+     * Converts {@link Article} list into generic {@link CardItem} model.
+     * <p>
+     * <br/>
+     * Check if we can use "adapter" or "factory" pattern to standardize this.
+     *
+     * @param articles List of articles.
+     * @return List of converted {@link CardItem}.
+     */
+    private List<CardItem> convertArticleToCardItems(final List<Article> articles) {
+        List<CardItem> cardItems = new ArrayList<>(articles.size());
+        for (Article result : articles) {
+            cardItems.add(CardItem.create(result));
+        }
+        return cardItems;
     }
 }
